requirements.txt
azure-identity
azure-keyvault-secrets
azure-ai-projects
fsspec
xarray
zarr
s3fs
adlfs
matplotlib>=3.7.0
azure-storage-blob>=12.19.0
azure-functions>=1.23.0
